#2155 create actions and reducer for fetching document data #2155 use redux action to fetch data #2155 divide DocumentList into a container/component #2155 render DocumentList with redux state data
# Conflicts:
#	src/components/app/DocumentList.js
#	src/utils/documentListHelper.js

# Conflicts:
#	src/components/app/DocumentList.js
#	src/components/table/Table.js
