+++ src/components/table/Table.js
@@ -80,6 +80,7 @@ class Table extends Component {
       tabId,
       isModal,
       hasIncluded,
+      page,
     } = this.props;
     const { selected, rows } = this.state;
     const selectedEqual = _.isEqual(prevState.selected, selected);
@@ -93,6 +94,11 @@ class Table extends Component {
     }
 
     if (rows && !_.isEqual(prevState.rows, rows)) {
+      this.setState({
+        collapsedRows: [],
+        collapsedParentsRows: [],
+      });
+
       if (isModal && !hasIncluded) {
         let firstRow = rows[0];
 
@@ -106,6 +112,11 @@ class Table extends Component {
           }
         }
       }
+    } else if (page !== prevProps.page) {
+      this.setState({
+        collapsedRows: [],
+        collapsedParentsRows: [],
+      });
     }
 
     if (mainTable && open) {
@@ -135,8 +146,15 @@ class Table extends Component {
         this.setState({ selected: [] });
       }
 
+      this.setState({
+        collapsedRows: [],
+        collapsedParentsRows: [],
+      });
+
       const firstLoad =
-        prevProps.rowData.get(`${tabId}`).size && rowData.get(`${tabId}`).size
+        prevProps.rowData.get(`${tabId}`) &&
+        prevProps.rowData.get(`${tabId}`).size &&
+        rowData.get(`${tabId}`).size
           ? false
           : true;
 
@@ -207,8 +225,9 @@ class Table extends Component {
     const { selected } = this.state;
     let rowsData = [];
 
-    if (indentSupported && rowData.get(`${tabId}`).size) {
+    if (indentSupported && rowData.get(`${tabId}`).length) {
       rowsData = getRowsData(rowData.get(`${tabId}`));
+
       let stateChange = {
         rows: rowsData,
         pendingInit: !rowsData,
@@ -225,6 +244,7 @@ class Table extends Component {
       this.setState(stateChange, () => {
         const { rows } = this.state;
         const firstRow = rows[0];
+
         let updatedParentsRows = [...this.state.collapsedParentsRows];
         let updatedRows = [...this.state.collapsedRows];
 
@@ -248,7 +268,7 @@ class Table extends Component {
 
         let mapCollapsed = [];
 
-        if (collapsible && rows && rows.length && selectFirst) {
+        if (collapsible && rows && rows.length) {
           rows.map((row) => {
             if (row.indent.length >= expandedDepth && row.includedDocuments) {
               mapCollapsed = mapCollapsed.concat(collapsedMap(row));
@@ -280,8 +300,8 @@ class Table extends Component {
       });
     } else {
       rowsData =
-        rowData.get(`${tabId}`) && rowData.get(`${tabId}`).size
-          ? rowData.get(`${tabId}`).toArray()
+        rowData.get(`${tabId}`) && rowData.get(`${tabId}`).length
+          ? rowData.get(`${tabId}`)
           : [];
 
       this.setState({
