+++ src/reducers/index.js
@@ -5,12 +5,14 @@ import listHandler from './listHandler';
 import menuHandler from './menuHandler';
 import windowHandler from './windowHandler';
 import pluginsHandler from './pluginsHandler';
+import dataHandler from './dataHandler';
 
 export default {
   appHandler,
   listHandler,
   menuHandler,
   windowHandler,
+  dataHandler,
   pluginsHandler,
   routing,
 };
